requirements.txt
amqp==2.3.2
ansible==2.6.4
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.4
boto==2.49.0
boto3==1.9.30
botocore==1.12.30
celery==4.2.1
cffi==1.11.5
cryptography==2.3.1
Django==1.11.15
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docutils==0.14
enum34==1.1.6
futures==3.2.0
idna==2.7
ipaddress==1.0.22
Jinja2==2.10
jmespath==0.9.3
kombu==4.2.1
MarkupSafe==1.0
paramiko==2.4.1
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.4
pycparser==2.18
PyJWT==1.7.1
PyNaCl==1.2.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
s3transfer==0.1.13
six==1.11.0
urllib3==1.24
vine==1.1.4
